Tweak blocks
- reinstated index/information pages
- folded old blocks into new structure
