+++ ons_alpha/core/blocks/__init__.py
@@ -1,5 +1,6 @@
-from .markup import HeadingBlock, TableBlock
+from .embeddable import DocumentBlock, ImageBlock
+from .markup import HeadingBlock, QuoteBlock, TableBlock, TypedTableBlock
 from .panels import PanelBlock
 
 
-__all__ = ["HeadingBlock", "PanelBlock", "TableBlock"]
+__all__ = ["DocumentBlock", "ImageBlock", "HeadingBlock", "QuoteBlock", "PanelBlock", "TableBlock", "TypedTableBlock"]
