+++ ons_alpha/core/blocks/embeddable.py
@@ -1,0 +1,21 @@
+from wagtail import blocks
+from wagtail.documents.blocks import DocumentChooserBlock
+from wagtail.images.blocks import ImageChooserBlock
+
+
+class ImageBlock(blocks.StructBlock):
+    image = ImageChooserBlock()
+    caption = blocks.CharBlock(required=False)
+
+    class Meta:
+        icon = "image"
+        template = "templates/components/streamfield/image_block.html"
+
+
+class DocumentBlock(blocks.StructBlock):
+    document = DocumentChooserBlock()
+    title = blocks.CharBlock(required=False)
+
+    class Meta:
+        icon = "doc-full-inverse"
+        template = "templates/components/streamfield/document_block.html"
