+++ ons_alpha/jinja2/templates/components/streamfield/heading_block.html
@@ -1,3 +1,3 @@
-<a href="#toc">Back to table of contents</a>
+{% if show_back_to_toc %}<a href="#toc">Back to table of contents</a>{% endif %}
 
 <h2 id="{{ value|slugify }}">{{ value }}</h2>
